requirements.txt
streamlit
streamlit-cookies-manager
python-dotenv
bcrypt
numpy
sqlite-utils
groq
pymongo
scikit-learn
fastapi
uvicorn
